+++ vite.config.ts
@@ -1,0 +1,8 @@
+import { cloudflare } from '@cloudflare/vite-plugin'
+import vue from '@vitejs/plugin-vue'
+import { defineConfig } from 'vite'
+import vuetify from 'vite-plugin-vuetify'
+
+export default defineConfig({
+  plugins: [vue(), vuetify({ autoImport: true }), cloudflare()],
+})